requirements.txt
fastapi
uvicorn
jinja2
requests
google-cloud-firestore
python-multipart